count speed correctly
We were currently using a special calculation for compound
meter, to make it play as iReal Pro says. However, this was
also throwing off the speed.
In this commit, we prioritize speed of the progression over
which beats the metronome should click on. Once we allow the
metronome to click at varying intervals, we will be able to
rewrite the failing tests for compound meters