don't hard reload when loading new song
We don't want to reload, because then the browser will wait for user input
to let us hijack the back button. If we don't reload, we stay within the same
window, and so we can control the back button.

It's important to make sure that we check window.history.state?.songScalerState
instead of window.history.state === null because we want to also hijack the
back button from use-query-params